requirements_base.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.8
decorator==5.1.1
distro==1.9.0
gpt4all==2.7.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==2.0.0
openai==1.14.3
outcome==1.3.0.post0
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.4.0
py==1.11.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pypdfium2==4.30.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
requests==2.32.3
retry==0.9.2
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
absl-py==2.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.3
blis==0.7.11
bs4==0.0.2
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
commonregex==1.5.4
confection==0.1.5
cramjam==2.8.3
cryptography==42.0.8
cssselect==1.2.0
cymem==2.0.8
dataclasses-json==0.6.7
datasets==2.20.0
decorator==5.1.1
deepeval==1.0.4
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
docx2txt==0.8
email_extractor==1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
evaluate==0.4.2
execnet==2.1.1
fastparquet==2024.5.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.15.4
flashgeotext==0.5.3
flashtext==2.7
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
fuzzywuzzy==0.18.0
geographiclib==2.0
geopy==2.4.1
geotext==0.4.0
googleapis-common-protos==1.63.2
gpt4all==2.7.0
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
importlib-metadata==7.0.0
iniconfig==2.0.0
jellyfish==1.1.0
jieba3k==0.35.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.32
langchain-openai==0.1.21
langchain-text-splitters==0.2.2
langcodes==3.4.0
langsmith==0.1.99
language_data==1.2.0
llmsherpa==0.1.4
locationtagger==0.0.1
loguru==0.7.2
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
openai==1.40.8
openpyxl==3.1.5
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
orjson==3.10.7
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.0
phonenumbers==8.13.42
pillow==10.4.0
pluggy==1.5.0
portalocker==2.10.1
preshed==3.0.9
probableparsing==0.0.1
protobuf==4.25.1
py==1.11.0
pyap==0.3.1
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycountry==24.6.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pypdfium2==4.30.0
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.2
pytest-repeat==0.9.3
pytest-xdist==3.6.1
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
ragas==0.1.14
redis==5.0.3
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
retry==0.9.2
rich==13.7.1
rouge_score==0.1.2
scikit-learn==1.5.1
scipy==1.14.0
selenium==4.22.0
sentry-sdk==2.13.0
setuptools==72.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.32
srsly==2.4.8
tabulate==0.9.0
tenacity==8.4.2
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
us==3.2.0
usaddress==0.5.10
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
win32-setctime==1.1.0
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.20.0
